CompetitiveRadar/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.41.0",
    "markdown>=3.9",
    "openai>=2.1.0",
]
